refactor: Move getMediaFilename to `egress` where it's only used there
