+++ pkg/upload/uploader.go
@@ -1,0 +1,8 @@
+package upload
+
+import "io"
+
+type Uploader interface {
+	// Key is a unique identifier for the file.
+	Upload(key string, body io.Reader) error
+}
